unwrap

+++ test/test_fori_loop_with_while_loop_simple_add_dispatch_in_torch.py
@@ -29,8 +29,8 @@ class WhileLoopTest(unittest.TestCase):
       return (x[0] +  x[0], x[1])
 
     device = xm.xla_device()
-    xx = torch.ones(1, dtype=torch.int64, device=device)
-    xy = torch.ones(10, dtype=torch.int64, device=device)
+    xx = torch.ones(1, dtype=torch.int32, device=device)
+    xy = torch.ones(10, dtype=torch.int32, device=device)
     res = while_loop(cond_fn, body_fn, (xx, xy))
     expected = _fake_while_loop(cond_fn, body_fn, x)
     self.assertEqual(expected, res)
